% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palletes.R
\name{adjust_colors}
\alias{adjust_colors}
\title{change saturation and brightness of given colors}
\usage{
adjust_colors(colors, saturation_factor = 1, brightness_factor = 1)
}
\arguments{
\item{colors}{vector of hex colors}

\item{saturation_factor}{factor}

\item{brightness_factor}{factor}
}
\value{
vector of adjusted colors
}
\description{
change saturation and brightness of given colors
}
\author{
Mandy Vogel
}
