% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aga_15.ref}
\alias{aga_15.ref}
\title{Parameters from recommendations of the German Adiposity Association (2015, AGA)}
\source{
"Kromeyer-Hauschild K, Moss A, Wabitsch M. Referenzwerte für den Body-Mass-Index für Kinder, Jugendliche und Erwachsene in Deutschland. Adipositas - Ursachen, Folgeerkrankungen, Therapie. 2015;09(3):123–7."
}
\usage{
aga_15.ref
}
\description{
Parameters from recommendations of the German Adiposity Association (2015, AGA)
}
