% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createlms.R
\name{aggregate_lms}
\alias{aggregate_lms}
\title{aggregate lms parameters}
\usage{
aggregate_lms(lms.list)
}
\arguments{
\item{lms.list}{list of parameter tables as returned by do_iterations()}
}
\value{
list of dataframes containing the aggregated parameters, each for every level of sex
}
\description{
aggregate lms parameters
}
\details{
function takes the lms part of the result from the do_iterations() function and returns
the mean parameters
}
\author{
Mandy Vogel
}
