% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{articles,refs-method}
\alias{articles,refs-method}
\title{Gives back references like articles or uris from an refs object, if \code{item} is given only the respective information will be extracted, if omitted information for all items will be returned}
\usage{
\S4method{articles}{refs}(object, item = NA)
}
\arguments{
\item{object}{object of class refs}

\item{item}{reference contained in \code{object}}

\item{...}{further arguments to \code{articles}}
}
\description{
Gives back references like articles or uris from an refs object, if \code{item} is given only the respective information will be extracted, if omitted information for all items will be returned
}

