% Generated by roxygen2 (4.0.1): do not edit by hand
\name{articles}
\alias{articles}
\title{give references like articles or uris}
\usage{
articles(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{further arguments to \code{articles}}
}
\description{
give references like articles or uris
}

