% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{belgium.ref}
\alias{belgium.ref}
\title{Parameters derived from Flandern population}
\source{
Roelants M, Hauspie R, Hoppenbrouwers K. References for growth and pubertal development from birth to 21 years in Flanders, Belgium. Annals of Human Biology. 2009 Dezember;36(6):680–94.
}
\usage{
belgium.ref
}
\description{
Parameters derived from Flandern population
}
