% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bone.ref}
\alias{bone.ref}
\title{Parameters for different bone parameters}
\source{
"Geserick M, Vogel M, Eckelt F, et al. Children and adolescents with obesity have reduced serum bone turnover markers and 25-hydroxyvitamin D but increased parathyroid hormone concentrations – Results derived from new pediatric reference ranges. Bone 2020;132:115124 and Weber et al. unpublishedfor VitD binding protein"
}
\usage{
bone.ref
}
\description{
Parameters for different bone parameters
}
