% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createlms.R
\name{calc_confints}
\alias{calc_confints}
\title{Calculate confidence intervals}
\usage{
calc_confints(lms.list, perc = c(2.5, 5, 50, 95, 97.5), level = 0.95,
  type = c("point"))
}
\arguments{
\item{lms.list}{lms part of the returned list of \code{\link{do_iterations}}}

\item{perc}{percentiles for which the confidence bands are calculated}

\item{level}{confidence level}

\item{type}{for now only point is a valid value}
}
\value{
list containing the respective confidence envelopes
}
\description{
Calculate confidence intervals
}
\details{
The function takes a lms list as returned by \code{\link{do_iterations}} and
calculates the confidence bands for a given set of percentiles using
\code{\link[boot]{envelope}} from the boot package
}
\author{
mandy
}
