% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdsvals.r
\name{calc_percent_excess}
\alias{calc_percent_excess}
\title{Calculate percentage relative to a given base percentile}
\usage{
calc_percent_excess(
  bmi = NULL,
  age = NULL,
  sex = NULL,
  ref.perc = 50,
  ref,
  item = "bmi",
  rownr = NULL
)
}
\arguments{
\item{bmi}{vector of bmi}

\item{age}{vector of age}

\item{sex}{vector of sex (coding "male" and "female") is assumend}

\item{ref.perc}{single value: reference percentile (0,100)}

\item{ref}{RefGroup object}

\item{item}{item within ref}

\item{rownr}{indicator of order}
}
\value{
vector containing values between 0 and 1
}
\description{
Calculate percentage relative to a given base percentile
}
\details{
The function calculates the percentage of a given bmi value relative to
a specific percentile
}
\author{
Mandy Vogel
}
