% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{cdc.ref}
\alias{cdc.ref}
\title{LMS Parameters for the Centers for Disease Control and Prevention 2000 Growth Charts, contains bmi, height, head cirumference, weight, weight for length,}
\source{
National health statitics reports 63.
}
\usage{
cdc.ref
}
\description{
LMS Parameters for the Centers for Disease Control and Prevention 2000 Growth Charts, contains bmi, height, head cirumference, weight, weight for length,
}

