% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cn.ref}
\alias{cn.ref}
\title{Parameters for height of normal weight and obese children from the CrescNet database
dependent on height}
\source{
"Kempf et al. In progress"
}
\usage{
cn.ref
}
\description{
Parameters for height of normal weight and obese children from the CrescNet database
dependent on height
}
