% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cole_lobstein.ref}
\alias{cole_lobstein.ref}
\title{Parameters for bmi from Cole and Lobstein, 2012}
\source{
Cole TJ, Lobstein T. Extended international (IOTF) body mass index cut-offs for thinness, overweight and obesity. Pediatric Obesity 2012;7(4):284–94.
}
\usage{
cole_lobstein.ref
}
\description{
Parameters for bmi from Cole and Lobstein, 2012
}
