% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colombia_sf.ref}
\alias{colombia_sf.ref}
\title{Parameters of skinfold measures derived from Colombian population}
\source{
Ramírez-Vélez, R. et al. Triceps and Subscapular Skinfold Thickness Percentiles and Cut-Offs for Overweight and Obesity in a Population-Based Sample of Schoolchildren and Adolescents in Bogota, Colombia. Nutrients 8, (2016).
}
\usage{
colombia_sf.ref
}
\description{
Parameters of skinfold measures derived from Colombian population
}
