% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{doyon_age.ref}
\alias{doyon_age.ref}
\title{Parameters for different carotid artery intima-media thickness and distensibility dependent on age}
\source{
"Doyon A, Kracht D, Bayazit AK, et al. Carotid artery intima-media thickness and distensibility in children and adolescents: reference values and role of body dimensions. Hypertension 2013;62(3):550–6"
}
\usage{
doyon_age.ref
}
\description{
Parameters for different carotid artery intima-media thickness and distensibility dependent on age
}
