% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dummy.refs}
\alias{dummy.refs}
\title{dummy}
\usage{
dummy.refs(ref, item)
}
\arguments{
\item{ref}{reference object}

\item{item}{item name specifiying the reference}
}
\value{
dummy output
}
\description{
dummy function to call a function of the gamlss.dist package
}
\details{
dummy function to call a function of the gamlss.dist package
}
\author{
Mandy Vogel
}
