% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ethiop.ref}
\alias{ethiop.ref}
\title{Parameters derived Ethiopian children}
\source{
Amare, E. B. et al. Reference Ranges for Head Circumference in Ethiopian Children 0--2 Years of Age. World Neurosurgery 84, 1566--1571.e2 (2015).
}
\usage{
ehtiop.ref
}
\description{
Parameters derived Ethiopian children
}
