% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ghouili_anthro.ref}
\alias{ghouili_anthro.ref}
\title{Parameters for height, weight, sitting height, etc from Ghouili, 2021}
\source{
Ghouili H, Ouerghi N, Boughalmi A, Dridi A, Rhibi F, Bouassida A. First growth reference curves for Tunisian children and adolescents. Archives de Pediatrie 2021;28(5):381–91
}
\usage{
ghouili_anthro.ref
}
\description{
Parameters for height, weight, sitting height, etc from Ghouili, 2021
}
