% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gomez_bmitmi.ref}
\alias{gomez_bmitmi.ref}
\title{Parameters for BMI and TMI from Gomez, 2021}
\source{
Gomez-Campos R, Vidal-Espinoza R, Marques de Moraes A, et al. Comparison of Anthropometric Indicators That Assess Nutritional Status From Infancy to Old Age and Proposal of Percentiles for a Regional Sample of Chile. Frontiers in Nutrition 2021 [cited 2022 May 28];8. Available from: https://www.frontiersin.org/article/10.3389/fnut.2021.657491
}
\usage{
gomez_bmitmi.ref
}
\description{
Parameters for BMI and TMI from Gomez, 2021
}
