% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iron.ref}
\alias{iron.ref}
\title{Parameters for iron-related blood parameters in children}
\source{
Rieger, K. et al. Reference intervals for iron-related blood parameters: results from a population-based cohort study (LIFE Child). LaboratoriumsMedizin 40, (2016).
}
\usage{
iron.ref
}
\description{
Parameters for iron-related blood parameters in children
}
