% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{japan_lab.ref}
\alias{japan_lab.ref}
\title{Parameters of serum insulin-like growth factor-I (IGF-I)}
\source{
Isojima, T., Shimatsu, A., Yokoya, S., Chihara, K., Tanaka, T., Hizuka, N., Teramoto, A., Tatsumi, K., Tachibana, K., Katsumata, N., Horikawa, R., 2012. Standardized centile curves and reference intervals of serum insulin-like growth factor-I (IGF-I) levels in a normal Japanese population using the LMS method. Endocrine Journal 59, 771–780. doi:10.1507/endocrj.EJ12-0110
}
\usage{
japan_lab.ref
}
\description{
Parameters of serum insulin-like growth factor-I (IGF-I)
}
