% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{japanese.ref}
\alias{japanese.ref}
\title{Parameters derived from Japanese children}
\source{
Inokuchi, M., Matsuo, N., Anzo, M., Takayama, J. I. & Hasegawa, T. Age-dependent percentile for waist circumference for Japanese children based on the 1992--1994 cross-sectional national survey data. Eur J Pediatr 166, 655--661 (2007)
}
\usage{
japanese.ref
}
\description{
Parameters derived from Japanese children
}
