% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kawel_boehm.ref}
\alias{kawel_boehm.ref}
\title{Parameters for Cardiovascular Magnetic Resonance}
\source{
"Kawel-Boehm N, Hetzel SJ, Ambale-Venkatesh B, et al. Reference ranges (“normal values”) for cardiovascular magnetic resonance (CMR) in adults and children: 2020 update. Journal of Cardiovascular Magnetic Resonance 2020;22(1):87."
}
\usage{
kawel_boehm.ref
}
\description{
Parameters for Cardiovascular Magnetic Resonance
}
