% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kiggs.ref}
\alias{kiggs.ref}
\title{LMS Parameters for German reference data (KiGGS, 2003-2006)
for height, weight, bmi, hip, whr, bodyfat, skinfold sum, triceps
skinfold, and waist circumference}
\source{
Referenzperzentile fuer anthropometrische Masszahlen und Blutdruck aus KiGGS 2003-2006, Robert Koch Institut, Germany
}
\usage{
kiggs.ref
}
\description{
LMS Parameters for German reference data (KiGGS, 2003-2006)
for height, weight, bmi, hip, whr, bodyfat, skinfold sum, triceps
skinfold, and waist circumference
}
