% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kiggs_bp.ref}
\alias{kiggs_bp.ref}
\title{Parameters derived from the German KiGGS cohort}
\source{
Neuhauser, H. K., Thamm, M., Ellert, U., Hense, H. W. & Rosario, A. S. Blood Pressure Percentiles by Age and Height from Nonoverweight Children and Adolescents in Germany. Pediatrics peds.2010-1290 (2011). doi:10.1542/peds.2010-1290.
}
\usage{
kiggs_bp.ref
}
\description{
Parameters derived from the German KiGGS cohort
}
\details{
contains 2-dimensional reference grid. Do not use with \code{\link{sds}} but \code{\link{sds_2d}}
}
