% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kiggs_lightness.ref}
\alias{kiggs_lightness.ref}
\title{Parameters regarding lightness and height-to-mass ratio}
\source{
Kliegl et al. in preparation
}
\usage{
kiggs_lightness.ref
}
\description{
Parameters regarding lightness and height-to-mass ratio
}
