% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kirk_bf.ref}
\alias{kirk_bf.ref}
\title{Parameters for fat and lean mass from Kirk, 2021}
\source{
Kirk B, Bani Hassan E, Brennan-Olsen S, et al. Body composition reference ranges in community-dwelling adults using dual-energy X-ray absorptiometry: the Australian Body Composition (ABC) Study. Journal of Cachexia, Sarcopenia and Muscle 2021;12(4):880–90.
}
\usage{
kirk_bf.ref
}
\description{
Parameters for fat and lean mass from Kirk, 2021
}
