% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kro.ref}
\alias{kro.ref}
\title{LMS Parameters for German reference data (Kromeyer Hauschild, 2001)
for height, weight, bmi, and waist circumference, including preterm
correction (Voigt)}
\source{
Perzentile fuer den Body-mass-Index fuer das Kindes- und Jugendalter unter Heranziehung verschiedener deutscher Stichproben, Monatsschrift Kinderheilkunde August 2001, Volume 149, Issue 8, pp 807-818; Fruehgeborenenkorrektur nach Voigt
}
\usage{
kro.ref
}
\description{
LMS Parameters for German reference data (Kromeyer Hauschild, 2001)
for height, weight, bmi, and waist circumference, including preterm
correction (Voigt)
}
