% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kro.ref15}
\alias{kro.ref15}
\title{LMS Parameters for German reference data (Kromeyer Hauschild, 2015)
for height, weight, bmi, and waist circumference, including preterm
correction (Voigt) and Kromyer-Hausschild 2015}
\source{
Perzentile fuer den Body-mass-Index fuer das Kindes- und Jugendalter unter Heranziehung verschiedener deutscher Stichproben, Monatsschrift Kinderheilkunde August 2001, Volume 149, Issue 8, pp 807-818; preterm adjustment Voigt; Adults: Kromeyer-Hauschild K, Moss A, Wabitsch M. Referenzwerte für den Body-Mass-Index für Kinder, Jugendliche und Erwachsene in Deutschland. Adipositas - Ursachen, Folgeerkrankungen, Therapie. 2015;09(3):123-127. doi:10.1055/s-0037-1618928
}
\usage{
kro.ref15
}
\description{
LMS Parameters for German reference data (Kromeyer Hauschild, 2015)
for height, weight, bmi, and waist circumference, including preterm
correction (Voigt) and Kromyer-Hausschild 2015
}
