% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leptin.ref}
\alias{leptin.ref}
\title{Parameters for leptin dependent on age/puberty and bmi}
\source{
Brandt-Hedunemann et al. In preparation
}
\usage{
leptin.ref
}
\description{
Parameters for leptin dependent on age/puberty and bmi
}
