% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liao_igf1.ref}
\alias{liao_igf1.ref}
\title{Parameters for IGF-1 from Liao, 2016}
\source{
Liao ZH, Yin QQ, Wan JX, He W, Ji W, Zhang LY, et al. Serum Insulin-like growth factor-1 levels of healthy adults in southern China. Endocr J. 2016 Dec 30;63(12):1081–6.
}
\usage{
liao_igf1.ref
}
\description{
Parameters for IGF-1 from Liao, 2016
}
