% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_ap.ref}
\alias{life_ap.ref}
\title{Parameters for alkaline phosphatase}
\source{
Strauch, J.-M., Vogel, M., Meigen, C., Ceglarek, U., Kratzsch, J., Willenberg, A., Kiess, W., 2023. Pediatric reference values of alkaline phosphatase: Analysis from a German population-based cohort and influence of anthropometric and blood parameters. Bone 174, 116809. https://doi.org/10.1016/j.bone.2023.116809
}
\usage{
life_ap.ref
}
\description{
Parameters for alkaline phosphatase
}
