% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_cbc.ref}
\alias{life_cbc.ref}
\title{Parameters for complete blood count}
\source{
Sommer et al. In preparation
}
\usage{
life_cbc.ref
}
\description{
Parameters for complete blood count
}
