% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_circ.ref}
\alias{life_circ.ref}
\title{Parameters for different circumferences and whr and whtr}
\source{
"Roennecke E, Vogel M, Bussler S, Grafe N, Jurkutat A, Schlingmann M, Koerner A, Kiess W. Age- and sex-related percentiles of skinfold thickness, waist and hip circumference, Waist-to- Hip Ratio and Waist-to-Height Ratio: Results from a population-based paediatric cohort in Germany (LIFE Child). Obesity Facts. 2019."
}
\usage{
life_circ.ref
}
\description{
Parameters for different circumferences and whr and whtr
}
