% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_cysc.ref}
\alias{life_cysc.ref}
\title{Parameters for different metabolom parameters from the LIFE Child cohort}
\source{
"Ziegelasch N, Vogel M, Müller E, et al. Cystatin C Serum Levels in Healthy Children Are Related to Age, Gender, and Pubertal Stage. Pediatr Nephrol 2019; 34: 449–57."
}
\usage{
life_cysc.ref
}
\description{
Parameters for different metabolom parameters from the LIFE Child cohort
}
