% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_fibroscan.ref}
\alias{life_fibroscan.ref}
\title{Parameters for fibroscan from the LIFE Child cohort}
\source{
Brunnert, L., Puasa, I.D., Garten, A., Penke, M., Gaul, S., Grafe, N., Karlas, T., Kiess, W., Flemming, G., Vogel, M., 2022. Pediatric percentiles for transient elastography measurements - effects of age, sex, weight status and pubertal stage. Front Endocrinol (Lausanne) 13, 1030809. https://doi.org/10.3389/fendo.2022.1030809
}
\usage{
life_fibroscan.ref
}
\description{
Parameters for fibroscan from the LIFE Child cohort
}
