% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_folb12.ref}
\alias{life_folb12.ref}
\title{Parameters regarding folate and cobalamine}
\source{
Kreusler, P., Vogel, M., Willenberg, A., Baber, R., Dietz, Y., Körner, A., Ceglarek, U., Kiess, W., 2021. Folate and Cobalamin Serum Levels in Healthy Children and Adolescents and Their Association with Age, Sex, Bmi and Socioeconomic Status. Nutrients 13. https://doi.org/10.3390/nu13020546
}
\usage{
life_folb12.ref
}
\description{
Parameters regarding folate and cobalamine
}
