% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_hba1c.ref}
\alias{life_hba1c.ref}
\title{Parameters for hba1c based on a child cohort a) including overweight and b) excluding overweight children}
\source{
"Hovestadt, I., Kiess, W., Lewien, C., Willenberg, A., Poulain, T., Meigen, C., Körner, A., Vogel, M., 2022. HbA1c Percentiles and the Association between BMI, Age, Gender, Puberty and HbA1c Levels in Healthy German Children and Adolescents. Pediatr Diabetes 23, 194–202. https://doi.org/10.1111/pedi.13297"

Hovestadt, I., Kiess, W., Lewien, C., Willenberg, A., Poulain, T., Meigen, C., Körner, A., Vogel, M., 2022. HbA1c Percentiles and the Association between BMI, Age, Gender, Puberty and HbA1c Levels in Healthy German Children and Adolescents. Pediatr Diabetes 23, 194–202. https://doi.org/10.1111/pedi.13297
}
\usage{
life_hba1c.ref

life_hba1c.ref
}
\description{
Parameters for hba1c based on a child cohort a) including overweight and b) excluding overweight children

Parameters regarding HbA1c
}
