% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_igf.ref}
\alias{life_igf.ref}
\title{IGF-I and IGF-BP3 from the LIFE Child cohort}
\source{
"Hoerenz C, Vogel M, Wirkner K. BMI and contraceptives affect new age-, sex-, and puberty-adjusted IGF-﻿I and IGFBP-﻿3 reference ranges across life span. JCEM 2022 (in (minor) revision)."
}
\usage{
life_igf.ref
}
\description{
IGF-I and IGF-BP3 from the LIFE Child cohort
}
