% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_liver.ref}
\alias{life_liver.ref}
\title{Parameters for serum liver enzymes}
\source{
Bussler et al, New pediatric percentiles of liver enzyme serum levels (ALT, AST, GGT): effects of age, sex, BMI and pubertal stage, Hepatology 2017
}
\usage{
life_liver.ref
}
\description{
Parameters for serum liver enzymes
}
