% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_oxyandrogene.ref}
\alias{life_oxyandrogene.ref}
\title{Parameters for oxyandrogenes}
\source{
Zeidler R, Wagner F, et al. in preparation
}
\usage{
life_oxyandrogene.ref
}
\description{
Parameters for oxyandrogenes
}
