% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_shbg.ref}
\alias{life_shbg.ref}
\title{Parameters for shbg and testosterone-shbg-ratio}
\source{
Duergen P. et al. in preparation
}
\usage{
life_shbg.ref
}
\description{
Parameters for shbg and testosterone-shbg-ratio
}
