% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_steroide.ref}
\alias{life_steroide.ref}
\title{Parameters for steroid hormones}
\source{
Bae, Y.J., Zeidler, R., Baber, R., Vogel, M., Wirkner, K., Loeffler, M., Ceglarek, U., Kiess, W., Koerner, A., Thiery, J., Kratzsch, J., 2019. Reference intervals of nine steroid hormones over the life-span analyzed by LC-MS/MS: Effect of age, gender, puberty, and oral contraceptives. J. Steroid Biochem. Mol. Biol. 193, 105409. https://doi.org/10.1016/j.jsbmb.2019.105409
}
\usage{
life_steroide.ref
}
\description{
Parameters for steroid hormones
}
