% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_stfr.ref}
\alias{life_stfr.ref}
\title{Parameters regarding soluble transferrin receptor}
\source{
Prenzel et al. In Preparation.
}
\usage{
life_stfr.ref
}
\description{
Parameters regarding soluble transferrin receptor
}
