% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_thyr.ref}
\alias{life_thyr.ref}
\title{Parameters for TSH, FT3, FT4 from the LIFE Child cohort}
\source{
"Surup H., Vogel M., Koerner A., Hiemisch A., Oelkers L., Willenberg A., Kiess W., Kratzsch J. (2021). BMI and puberty have to be included into the interpretation of TSH, FT3 and FT4 measurements by new pediatric reference intervals. THYROID."
}
\usage{
life_thyr.ref
}
\description{
Parameters for TSH, FT3, FT4 from the LIFE Child cohort
}
