% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linden_heart.ref}
\alias{linden_heart.ref}
\title{Parameters for left artrial volumne Linden et al,}
\source{
Linden, K. et al. (2019) Left Atrial Volumes and Phasic Function in Healthy Children: Reference Values Using Real-Time Three-Dimensional Echocardiography, Journal of the American Society of Echocardiography, 32(8), pp. 1036-1045.e9. doi:10.1016/j.echo.2019.03.018.
}
\usage{
linden_heart.ref
}
\description{
Parameters for left artrial volumne Linden et al,
}
