% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lipids.ref}
\alias{lipids.ref}
\title{Parameters for serum lipids in children}
\source{
Dathan-Stumpf, A. et al. Pediatric reference data of serum lipids and prevalence of dyslipidemia: Results from a population-based cohort in Germany. Clinical Biochemistry 49, 740–749 (2016).
}
\usage{
lipids.ref
}
\description{
Parameters for serum lipids in children
}
