% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_percentile_tab}
\alias{make_percentile_tab}
\title{calculate raw values}
\usage{
make_percentile_tab(refs, item, perc = c(1, 5, 50, 95, 99), stack = F, sex,
  age)
}
\arguments{
\item{refs}{Refgroup object}

\item{item}{name of the measurement item}

\item{perc}{vector of percentiles to be calculated}

\item{stack}{wether or not the data should be stacked, stacked data
would most possibly be used in ggplot2}

\item{sex}{name of the sex variable (character) if different from sex, ignored}

\item{age}{name of the age variable (character) if different from age, ignored}
}
\value{
data frame either with the different percentiles as columns
or, if stacked, as data frame with four columns: age, sex, variable, value
}
\description{
Calculate raw values for percentile curve
}
\details{
calculates quantile values for given RefGroup and given
percentiles
}
\author{
Mandy Vogel
}
