% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridits.R
\name{mean_ridits}
\alias{mean_ridits}
\title{mean ridits for vectors of paired observations}
\usage{
mean_ridits(v, w, maxrank = max(v), minrank = NULL)
}
\arguments{
\item{v}{vector 1}

\item{w}{vector 2}

\item{maxrank}{max possible rank}

\item{minrank}{min possible rank}
}
\value{
2-element named vector
}
\description{
mean ridits for vectors of paired observations
}
\author{
Mandy Vogel
}
