% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mock_df}
\alias{mock_df}
\title{Mock a data frame}
\usage{
mock_df(ref, item, n = 1000)
}
\arguments{
\item{ref}{a valid RefGroup object}

\item{item}{a valid imte present in ref}

\item{n}{how many values should be created}
}
\value{
data frame containing a age, sex, and value column
}
\description{
mock values for a given reference
}
\details{
mock values for a given reference
}
\author{
mandy
}
