% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mock_value}
\alias{mock_value}
\title{Mock Value}
\usage{
mock_value(ref, item, sex = c("male", "female"), age)
}
\arguments{
\item{ref}{a valid RefGroup object}

\item{item}{a valid item present in ref}

\item{sex}{character male or female}

\item{age}{numeric age value}
}
\value{
a random value from the conditional distribution (conditionally on age and sex)
}
\description{
mock a value for a given reference
}
\details{
the function creates a random value for a given age and sex value and a given reference
}
\author{
mandy
}
