% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{mock_values}
\alias{mock_values}
\title{Mock Values}
\usage{
mock_values(df, sex, age, ref, item)
}
\arguments{
\item{df}{data frame containing the age and sex}

\item{sex}{name of the sex variable}

\item{age}{name of the age variable}

\item{ref}{a valid RefGroup object}

\item{item}{a valid imte present in ref}
}
\value{
data frame containing the additional column with random numbers
}
\description{
mock values for a given reference, given age and given sex
}
\details{
the function creates random values for given age and sex values and a given reference
}
\author{
mandy
}
