% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{momo.ref}
\alias{momo.ref}
\title{Parameters for the German MoMo study (sports test)}
\source{
"Niessner C, Utesch T, Oriwol D, et al. Representative Percentile Curves of Physical Fitness From Early Childhood to Early Adulthood: The MoMo Study. Front Public Health 2020;8. Available from: https://www.frontiersin.org/articles/10.3389/fpubh.2020.00458/full?report=reader"
}
\usage{
momo.ref
}
\description{
Parameters for the German MoMo study (sports test)
}
