% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motor.ref}
\alias{motor.ref}
\title{Parameters for 5 subtests of the KiGGS Motorik Module}
\source{
"Sobek et al. In progress"
}
\usage{
motor.ref
}
\description{
Parameters for 5 subtests of the KiGGS Motorik Module
}
