% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nl3.ref}
\alias{nl3.ref}
\title{Parameters of skinfold measures derived from Colombian population}
\source{
Fredriks, A. M. et al. Continuing positive secular growth change in The Netherlands 1955–1997. Pediatric research 47, 316–323 (2000).

Fredriks, A.M., van Buuren, S., Wit, J.M., Verloove-Vanhorick, S.P., 2000. Body index measurements in 1996–7 compared with 1980. Archives of disease in childhood 82, 107–112.

https://cran.r-project.org/package=AGD
}
\usage{
nl3.ref
}
\description{
Parameters of skinfold measures derived from Colombian population
}
