% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nl4.ref}
\alias{nl4.ref}
\title{Parameters derived from the 4th Dutch growth study}
\source{
Fredriks, A. M. et al. Nationwide age references for sitting height, leg length, and sitting height/height ratio, and their diagnostic value for disproportionate growth disorders. Archives of Disease in Childhood 90, 807--812 (2005);
Fredriks, A. M. et al. Height, weight, body mass index and pubertal development references for children of Moroccan origin in The Netherlands. Acta Paediatr. 93, 817--824 (2004);
Fredriks, A. M. et al. Continuing positive secular growth change in The Netherlands 1955--1997. Pediatric research 47, 316--323 (2000);
Fredriks, A. M. et al. Height, weight, body mass index and pubertal development reference values for children of Turkish origin in the Netherlands. Eur. J. Pediatr. 162, 788--793 (2003); Fredriks, A. M., van Buuren, S., Wit, J. M. & Verloove-Vanhorick, S. P. Body index measurements in 1996--7 compared with 1980. Archives of disease in childhood 82, 107--112 (2000);
R package: AGD, Stef van Buuren, http://www.stefvanbuuren.nl/
}
\usage{
nl4.ref
}
\description{
Parameters derived from the 4th Dutch growth study
}
