% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ofenheimer_bf.ref}
\alias{ofenheimer_bf.ref}
\title{Parameters for fat and lean mass from Ofenheimer et al., 2020}
\source{
Ofenheimer A, Breyer-Kohansal R, Hartl S, et al. Reference values of body composition parameters and visceral adipose tissue (VAT) by DXA in adults aged 18–81 years—results from the LEAD cohort. European Journal of Clinical Nutrition 2020;74(8):1181–91.
}
\usage{
ofenheimer_bf.ref
}
\description{
Parameters for fat and lean mass from Ofenheimer et al., 2020
}
