% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refsclass.r
\docType{class}
\name{ParTab-class}
\alias{ParTab-class}
\title{Table of references}
\description{
Reference tables
}
\section{Slots}{

\describe{
\item{\code{item}}{identifier of the item}

\item{\code{dist}}{named list which contains the distribution which was used in fitting 
the references. One entry for male and one for female}
}}

