% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preterm.ref}
\alias{preterm.ref}
\title{Parameters Preterm and  Intrauterine}
\source{
Olsen, I.E., Lawson, M.L., Ferguson, A.N., Cantrell, R., Grabich, S.C., Zemel, B.S., Clark, R.H., 2015. BMI Curves for Preterm Infants. PEDIATRICS 135, e572–e581. doi:10.1542/peds.2014-2777
}
\usage{
preterm.ref
}
\description{
Parameters Preterm and  Intrauterine
}
