% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridits.R
\name{props}
\alias{props}
\title{propabilities}
\usage{
props(v, maxrank = max(v), minrank = NULL)
}
\arguments{
\item{v}{vector of ranks}

\item{maxrank}{max possible rank}

\item{minrank}{min possible rank}
}
\value{
1-dim contingency table
}
\description{
propabilities
}
\author{
Mandy Vogel
}
