% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palletes.R
\name{prs_pal}
\alias{prs_pal}
\title{color palette Pediatric research}
\usage{
prs_pal(pal = "orig", primary = "blue", other = "violet", direction = 1)
}
\arguments{
\item{pal}{orig, sat or bright}

\item{primary}{first colour for two-color scales, must be one of the scale colors}

\item{other}{second colour for two-color scales}

\item{direction}{if -1 one order of colours is reversed}
}
\value{
function of n
}
\description{
color palette Pediatric research
}
\author{
Mandy Vogel
}
