% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refsclass.r
\docType{class}
\name{RefGroup-class}
\alias{RefGroup-class}
\title{Class of references}
\description{
Container for reference tables
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the reference group}

\item{\code{refs}}{List of references, each reference refers to one item and contains
independent variable age, and the parameter values for both genders}

\item{\code{citations}}{information about the sources of the references}

\item{\code{info}}{additional infos regarding the references}
}}

\examples{
data(kiggs.ref)
print(kiggs.ref)
data(ukwho.ref)
print(ukwho.ref)
data(who.ref)
print(who.ref)
}
\author{
Mandy Vogel
}
