% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{refs-class}
\alias{refs-class}
\title{Class of references}
\description{
Container for reference tables
}
\section{Slots}{

\describe{
\item{\code{name}}{of the reference group}

\item{\code{ref}}{List of references, each reference refers to one item and contains age, l, m, and s values for both genders}

\item{\code{citation}}{information about the sources of the references}
}}
\examples{
data(kiggs.ref)
print(kiggs.ref)
data(ukwho.ref)
print(ukwho.ref)
data(who.ref)
print(who.ref)
}

