% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridits.R
\name{ridits.from.ranks}
\alias{ridits.from.ranks}
\title{ridits from ranks}
\usage{
ridits.from.ranks(v, maxrank = max(v), minrank = NULL)
}
\arguments{
\item{v}{vector of ranks}

\item{maxrank}{min possible rank}

\item{minrank}{max possible rank}
}
\value{
vector of ridits
}
\description{
ridits from ranks
}
\author{
Mandy Vogel
}
