% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ripka_bf.ref}
\alias{ripka_bf.ref}
\title{Parameters for fat and lean mass from Ripka et al, 2020}
\source{
Ripka, W.L. et al. (2020) Lean mass reference curves in adolescents using dual-energy x-ray absorptiometry (DXA), PLOS ONE, 15(2), p. e0228646. doi:10.1371/journal.pone.0228646.
}
\usage{
ripka_bf.ref
}
\description{
Parameters for fat and lean mass from Ripka et al, 2020
}
