% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palletes.R
\name{scale_fill_prs}
\alias{scale_fill_prs}
\title{discrete ggplot colour scale of PRS colors}
\usage{
scale_fill_prs(
  pal = "orig",
  primary = "blue",
  other = "violet",
  direction = 1,
  ...
)
}
\arguments{
\item{pal}{orig, sat, or bright}

\item{primary}{primary colour of two}

\item{other}{the other colour}

\item{direction}{if -1 one order of colours is reversed}

\item{...}{further arguments}
}
\value{
colour scale for use with ggplot()
}
\description{
discrete ggplot fill scale
}
\author{
Mandy Vogel
}
