% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palletes.R
\name{scale_uchu}
\alias{scale_uchu}
\alias{scale_colour_uchu}
\alias{scale_fill_uchu}
\title{uchu colour scale}
\usage{
scale_colour_uchu(
  pal = c("gray", "red", "pink", "purple", "blue", "green", "yellow", "orange",
    "general"),
  ...
)

scale_fill_uchu(
  pal = c("gray", "red", "pink", "purple", "blue", "green", "yellow", "orange",
    "general"),
  ...
)
}
\arguments{
\item{pal}{colour palette}

\item{...}{further arguments to scale}
}
\value{
ggplot colour scale
}
\description{
Uchu colour scale
}
\details{
provides the uchu colour scale as ggplot2 colour scale (https://github.com/NeverCease/uchu)
}
\author{
Mandy Vogel
}
