% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schafmeyer_leg.ref}
\alias{schafmeyer_leg.ref}
\title{Parameters for bone mineral content/density, fm, lm (lower limbs) from Schafmeyer, 2022}
\source{
Schafmeyer L, Linden T, Sill H, Rehberg M, Schoenau E, Duran I. Pediatric Reference Centiles of Bone Mineral Density and Body Composition of Lower Limbs. Journal of Clinical Densitometry. 2022 Jan;25(1):73–80.
}
\usage{
schafmeyer_leg.ref
}
\description{
Parameters for bone mineral content/density, fm, lm (lower limbs) from Schafmeyer, 2022
}
