% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{sds,refs-method}
\alias{sds,refs-method}
\title{calculates standard deviation scores according to a given reference}
\usage{
\S4method{sds}{refs}(object, item, sex, age, value)
}
\arguments{
\item{object}{\code{\linkS4class{refs}} object}

\item{item}{reference contained in \code{object}}

\item{sex}{gender}

\item{age}{age}

\item{value}{numeric value which should transform into SDscore}
}
\description{
calculates standard deviation scores according to a given reference
}
\examples{
data(who.ref)
sds(who.ref,item="height",sex="female",age=0,value=c(50,51,52,53))
sds(who.ref,item="weight",sex="female",age=0,value=c(2.5,3,3.5))
sds(who.ref,item="bmi",sex="male",age=c(10,11,12),value=c(14.5,15,19.5))
}

