% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sds}
\alias{sds}
\title{calculates standard deviation scores according to a given reference}
\usage{
sds(object, ...)
}
\arguments{
\item{object}{an \code{\linkS4class{refs}} object}

\item{...}{further arguments to \code{sds}}
}
\description{
calculates standard deviation scores according to a given reference
}

