% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createlms.R
\name{select_meas}
\alias{select_meas}
\title{choose one measurement per subject}
\usage{
select_meas(data, subject = "subject", prop = 1, verbose = F)
}
\arguments{
\item{data}{dataframe as returned by prepare data}

\item{subject}{name of the column containing the subject identifier}

\item{prop}{optional - proportion of measurements to sample}

\item{verbose}{if TRUE information about sample size is printed out}
}
\value{
dataframe containing the sampled rows
}
\description{
Choose one measurement per subject
}
\details{
function samples one measurement per subject, if prop < 1 additional
a prop*100 percent will be sampled from the measurements
the function is called inside \code{\link{do_iterations}} and may not called directly
}
\author{
Mandy Vogel
}
