% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{show,refs-method}
\alias{show,refs-method}
\title{Shows \code{items} and additional information of an \code{refs} object}
\usage{
\S4method{show}{refs}(object)
}
\arguments{
\item{object}{\code{refs} object}
}
\description{
Shows \code{items} and additional information of an \code{refs} object
}

