% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refsclass.r
\docType{methods}
\name{show,ParTab-method}
\alias{show,ParTab-method}
\title{class ParTab}
\usage{
\S4method{show}{ParTab}(object)
}
\arguments{
\item{object}{object of calss ParTab}
}
\value{
print information about the respective reference table
}
\description{
show method for ParTab
}
\details{
show method for ParTab
}
\author{
Mandy Vogel
}
