% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refsclass.r
\docType{methods}
\name{show,RefGroup-method}
\alias{show,RefGroup-method}
\title{class RefGroup}
\usage{
\S4method{show}{RefGroup}(object)
}
\arguments{
\item{object}{object of class RefGroup}
}
\value{
prints information about age range, citations, etc.
}
\description{
show method for RefGroup
}
\details{
show method for RefGroup
}
\author{
Mandy Vogel
}
