% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turkish.ref}
\alias{turkish.ref}
\title{Parameters derived from Turkish children}
\source{
Hatipoglu, N. et al. Waist circumference percentiles for 7- to 17-year-old Turkish children and adolescents. Eur J Pediatr 167, 383--389 (2008);Bundak, R. et al. Body mass index references for Turkish children. Acta Pædiatrica 95, 194--198 (2006).
}
\usage{
turkish.ref
}
\description{
Parameters derived from Turkish children
}
