% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uk1990.ref}
\alias{uk1990.ref}
\title{Parameters from the 1990 UK growth study}
\source{
Cole, T.J., Freeman, J.V., Preece, M.A., 1998. British 1990 growth reference centiles for weight, height, body mass index and head circumference fitted by maximum penalized likelihood. Statistics in medicine 17, 407–429.

Cole, T.J., Freeman, J.V., Preece, M.A., 1995. Body mass index reference curves for the UK, 1990. Archives of disease in childhood 73, 25–29.
}
\usage{
uk1990.ref
}
\description{
Parameters from the 1990 UK growth study
}
