% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukwho.ref}
\alias{ukwho.ref}
\title{LMS Parameters for UK-WHO growth charts for height,
weight, bmi, head circumference}
\source{
Wright, Charlotte M., et a,Practice pointer: Using the new UK-WHO growth charts. British Medical Journal 340.c1140 (2010): 647-650.Preterm British 1990, 0-4 WHO2006, 4-18 British1990
}
\usage{
ukwho.ref
}
\description{
LMS Parameters for UK-WHO growth charts for height,
weight, bmi, head circumference
}
