% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us.ref}
\alias{us.ref}
\title{Parameters derived from US children (additional to the cdc.ref)}
\source{
Sharma, A. K., Metzger, D. L., Daymont, C., Hadjiyannakis, S. & Rodd, C. J. LMS tables for waist-circumference and waist-height ratio Z-scores in children aged 5-19 y in NHANES III: association with cardio-metabolic risks. Pediatric research (2015)
}
\usage{
us.ref
}
\description{
Parameters derived from US children (additional to the cdc.ref)
}
