% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valencia_nc.ref}
\alias{valencia_nc.ref}
\title{Parameters for neck circumference from Valencia-Sosa, 2021}
\source{
Valencia-Sosa E, Chavez-Palencia C, Vallarta-Robledo JR, et al. Percentile Reference Values for the Neck Circumference of Mexican Children. Children 2021;8(5):407.
}
\usage{
valencia_nc.ref
}
\description{
Parameters for neck circumference from Valencia-Sosa, 2021
}
