% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{wormplot2d_gg}
\alias{wormplot2d_gg}
\title{Worm Plot ggplot version}
\usage{
wormplot2d_gg(
  m = NULL,
  residuals = NULL,
  age = NA,
  x2 = NA,
  name.x2,
  n.inter.age = 1,
  n.inter.x2 = 1,
  y.limits = c(-1, 1)
)
}
\arguments{
\item{m}{a gamlss model}

\item{residuals}{nlormalized quantile residuals}

\item{age}{numeric vector of ages}

\item{x2}{numeric vector of second covariate}

\item{name.x2}{name of x2 for use in graphics}

\item{n.inter.age}{number of age intervals or cut points}

\item{n.inter.x2}{number of x2 intervals or cut points}

\item{y.limits}{limits of the y-axis}
}
\value{
ggplot object
}
\description{
Worm plot ggplot version, 2 covariates
}
\details{
creates a wormplot for a gamlss model or a given vector of
normalized quantile residuals, either for all residuals or
grouped by age intervals
}
