% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{wormplot_gg}
\alias{wormplot_gg}
\title{Worm Plot ggplot version}
\usage{
wormplot_gg(m = NULL, residuals = NULL, age = NA, n.inter = 1,
  y.limits = c(-1, 1))
}
\arguments{
\item{m}{a gamlss model}

\item{residuals}{normalized quantile residuals}

\item{age}{numeric vector of ages}

\item{n.inter}{number of age intervals or cut points}

\item{y.limits}{limits of the y-axis}
}
\value{
ggplot object
}
\description{
Worm plot ggplot version
}
\details{
creates a wormplot for a gamlss model or a given vector of
normalized quantile residuals, either for all residuals or
grouped by age intervals
}
