% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bp_wuehl_height.ref}
\alias{bp_wuehl_height.ref}
\title{Parameters from Wuehl et al. blood pressure reference values Germany according to height
from version 0.7.3 unplausible values are replaced by interpolated ones. For the original values
check out earlier versions}
\source{
"Wuehl E, Witte K, Soergel M, Mehls O, Schaefer F, Hypertension  for the GWG on P. Distribution of 24-h ambulatory blood pressure in children: normalized reference values and role of body dimensions. Journal of Hypertension. 2002 Oct;20(10):1995.", implausible values were replaced by interpolated ones from package version 0.7.4
}
\usage{
bp_wuehl_height.ref
}
\description{
Parameters from Wuehl et al. blood pressure reference values Germany according to height
from version 0.7.3 unplausible values are replaced by interpolated ones. For the original values
check out earlier versions
}
