% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createlms.R
\name{fit_gamlss}
\alias{fit_gamlss}
\title{fit lms}
\usage{
fit_gamlss(data, age.min = 0.25, age.max = 18, age.int = 1/12,
  dist = "BCCGo", mu.df = 4, sigma.df = 3, nu.df = 2, tau.df = 2,
  value)
}
\arguments{
\item{data}{dataframe as return by select_meas()}

\item{age.min}{lower bound of age}

\item{age.max}{upper bound of age}

\item{age.int}{stepwidth of the age variable}

\item{dist}{distribution used for the fitting process, has to be one of BCCGo, BCPEo, BCTo as they are accepted by lms()}

\item{mu.df}{degree of freedem location parameter}

\item{sigma.df}{degree of freedem spread parameter}

\item{nu.df}{degree of freedem skewness parameter}

\item{tau.df}{degree of freedem kurtosis parameter}

\item{value}{names of the value variable (character) if different from value, ignored}
}
\value{
list containing a dataframe of the fitted lms parameter at the given age points and the fitted model
}
\description{
fit lms
}
\details{
wrapper around the \code{\link[gamlss]{lms}} function in the gamlss package
returns the fitted lms-parameter at given age points
the function is called inside \code{\link{do_iterations}} and may not called directly
}
\author{
Mandy Vogel
}
