% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_fibroscan.ref}
\alias{life_fibroscan.ref}
\title{Parameters for fibroscan from the LIFE Child cohort}
\source{
"preliminary reference values cap med and e med from fibroscan in the life child study. Publication (Puasa et al) in preparation."
}
\usage{
life_fibroscan.ref
}
\description{
Parameters for fibroscan from the LIFE Child cohort
}
