% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_heart.ref}
\alias{life_heart.ref}
\title{hs-Troponin T and NT-proBNP from the LIFE Child cohort}
\source{
"Kiess A, Green J, Willenberg A, et al. Age-dependent reference values for hs-Troponin T and NT-proBNP and determining factors in a cohort of healthy children (The LIFE child study). Pediatric Cardiology 2022. Accepted"
}
\usage{
life_heart.ref
}
\description{
hs-Troponin T and NT-proBNP from the LIFE Child cohort
}
