% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{life_thyr.ref}
\alias{life_thyr.ref}
\title{Parameters for TSH, FT3, FT4 from the LIFE Child cohort
dependent on height}
\source{
"Surup et al. In progress"
}
\usage{
life_thyr.ref
}
\description{
Parameters for TSH, FT3, FT4 from the LIFE Child cohort
dependent on height
}
