% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refsclass.r
\name{sds}
\alias{sds}
\title{Calculate SDS Values}
\usage{
sds(value, age, sex, item, ref, type = "SDS", male = "male",
  female = "female")
}
\arguments{
\item{value}{vector of measurement values}

\item{age}{vector of age values}

\item{sex}{vector of sex}

\item{item}{name of the item e.g. "height"}

\item{ref}{RefGroup object}

\item{type}{"SDS" or "perc"}

\item{male}{coding of sex for male}

\item{female}{coding of sex for male}
}
\value{
vector containing SDS or percentile values
}
\description{
Calculate SDS values
}
\details{
The function takes a vector of measurement values, and of age and of sex
and a RefGroup object as arguments. It calculates the sds or percentile
values.
}
\author{
Mandy Vogel
}
