% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createlms.R
\name{select_fams}
\alias{select_fams}
\title{select families}
\usage{
select_fams(data, prop = 0.75, group, verbose = F)
}
\arguments{
\item{data}{dataframe as returned by prepare data}

\item{prop}{proportion of families to be sampled}

\item{group}{name of the group variable (character) if not "group", ignored}

\item{verbose}{if TRUE information about sample size is printed out}
}
\value{
dataframe containing only prop.fam percent the families in data
}
\description{
Select groups (families)
}
\details{
function selects a given proportion of groups/families from the data
if no grouping variable is given the original data set is returned
function is called inside \code{\link{do_iterations}} and may not called directly
}
\author{
Mandy Vogel
}
