% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who.ref}
\alias{who.ref}
\title{LMS Parameters for UK-WHO growth charts for height, weight, bmi,
head circumference,arm mid upper arm circumference, subscapular
and triceps skinfold, weight for height}
\source{
de Onis, M., Onyango, A., Borghi, E., Siyam, A., Blossner, M., & Lutter, C. (2012). Worldwide implementation of the WHO child growth standards. Public Health Nutr, 12, 1-8.
}
\usage{
who.ref
}
\description{
LMS Parameters for UK-WHO growth charts for height, weight, bmi,
head circumference,arm mid upper arm circumference, subscapular
and triceps skinfold, weight for height
}
